♻️ Extract test scenarios from auth.ts into scenarios.ts
Separate concerns: auth.ts now only handles accounts, login helpers,
and browser contexts. Test scenario definitions (PickBanBehavior,
TestScenario, testScenarios, testPairs) moved to scenarios.ts.
